+++ database/migrations/2025_06_19_072804_create_districts_table.php
@@ -8,31 +8,44 @@ return new class extends Migration
 {
     public function up()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
-        $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
-        $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
+        if (config('indonesia.data_location.district')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
+            $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
+            $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
 
-        $cityTable = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kabupaten' : 'cities');
-        $cityCode = config('indonesia.pattern') === 'ID' ? 'kode_kabupaten' : 'city_code';
+            if (config('indonesia.data_location.city')) {
+                $cityTable = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kabupaten' : 'cities');
+                $cityCode = config('indonesia.pattern') === 'ID' ? 'kode_kabupaten' : 'city_code';
+            } else {
+                $cityTable = null;
+                $cityCode = null;
+            }
 
-        Schema::create($tableName, function (Blueprint $table) use ($code, $name, $cityCode, $cityTable) {
-            $table->bigIncrements('id');
-            $table->char($code, 7)->unique();
-            $table->char($cityCode, 4);
-            $table->string($name, 255);
-            $table->text('meta')->nullable();
-            $table->timestamps();
+            Schema::create($tableName, function (Blueprint $table) use ($code, $name, $cityCode, $cityTable) {
+                $table->bigIncrements('id');
+                $table->char($code, 7)->unique();
+                if (config('indonesia.data_location.city')) {
+                    $table->char($cityCode, 4);
+                }
+                $table->string($name, 255);
+                $table->text('meta')->nullable();
+                $table->timestamps();
 
-            $table->foreign($cityCode)
-                ->references($code)
-                ->on($cityTable)
-                ->onUpdate('cascade')->onDelete('restrict');
-        });
+                if (config('indonesia.data_location.city')) {
+                    $table->foreign($cityCode)
+                        ->references($code)
+                        ->on($cityTable)
+                        ->onUpdate('cascade')->onDelete('restrict');
+                }   
+            });
+        }
     }
 
     public function down()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
-        Schema::drop($tableName);
+        if (config('indonesia.data_location.district')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
+            Schema::drop($tableName);
+        }
     }
 };
